requirements.txt
flask>=1.0.2
flask-sqlalchemy>=2.3.0
psycopg2-binary>=2.8.4
python-dotenv>=0.20
wtforms>=3.0.0
Werkzeug>=2.0
Flask-WTF>=1.0.0
pymysql>=1.0
Flask-Login>=0.5